+++ src/marshmallow/utils.py
@@ -6,7 +6,6 @@ import datetime as dt
 import inspect
 import typing
 from collections.abc import Mapping
-from email.utils import format_datetime, parsedate_to_datetime
 
 EXCLUDE = "exclude"
 INCLUDE = "include"
@@ -49,44 +48,11 @@ def is_collection(obj) -> bool:
     return is_iterable_but_not_string(obj) and not isinstance(obj, Mapping)
 
 
-def is_keyed_tuple(obj) -> bool:
-    """Return True if ``obj`` has keyed tuple behavior, such as
-    namedtuples or SQLAlchemy's KeyedTuples.
-    """
-    return isinstance(obj, tuple) and hasattr(obj, "_fields")
-
-
 # https://stackoverflow.com/a/27596917
 def is_aware(datetime: dt.datetime) -> bool:
     return (
         datetime.tzinfo is not None and datetime.tzinfo.utcoffset(datetime) is not None
     )
-
-
-def from_rfc(datestring: str) -> dt.datetime:
-    """Parse a RFC822-formatted datetime string and return a datetime object.
-
-    https://stackoverflow.com/questions/885015/how-to-parse-a-rfc-2822-date-time-into-a-python-datetime  # noqa: B950
-    """
-    return parsedate_to_datetime(datestring)
-
-
-def rfcformat(datetime: dt.datetime) -> str:
-    """Return the RFC822-formatted representation of a datetime object.
-
-    :param datetime: The datetime.
-    """
-    return format_datetime(datetime)
-
-
-def get_fixed_timezone(offset: int | float | dt.timedelta) -> dt.timezone:
-    """Return a tzinfo instance with a fixed offset from UTC."""
-    if isinstance(offset, dt.timedelta):
-        offset = offset.total_seconds() // 60
-    sign = "-" if offset < 0 else "+"
-    hhmm = "{:02d}{:02d}".format(*divmod(abs(offset), 60))
-    name = sign + hhmm
-    return dt.timezone(dt.timedelta(minutes=offset), name)
 
 
 def from_timestamp(value: typing.Any) -> dt.datetime: